+++ R/lazy-methods.R
@@ -22,7 +22,8 @@ format.lazyraster <- function(x, ...) {
   ex <- to_xy_minmax(object)
   windowdescription <- if (is.null(object$window$windowextent)) "<whole extent>" else paste(format(object$window$windowextent, nsmall = 4), collapse = ", ")
   windowindex <- if (is.null(object$window$window)) "<->" else paste(as.integer(object$window$window), collapse = ", ")
-
+  windowdim <-   if (is.null(object$window$window)) "(full)" else paste(as.integer(c(diff(object$window$window[1:2]),
+                                                                                     diff(object$window$window[3:4]))) + 1L, collapse = ", ")
   x <- list(   classname = "LazyRaster",
                dimensions = object$info$dimXY,
                resolution = abs(object$info$geotransform[c(2, 6)]),
@@ -34,13 +35,15 @@ format.lazyraster <- function(x, ...) {
                window = windowindex)
 
 
-  fmt <- list(sprintf("class         : %s\n", x$classname),
-              sprintf("dimensions    : %s (nrow, ncol)\n", paste(x$dimension[2:1], collapse = ", ")),
-              sprintf("resolution    : %s (x, y)\n", paste(format(x$resolution, nsmall = 4), collapse = ", ")),
-              sprintf("extent        : %s (xmin, xmax, ymin, ymax)\n", paste(format(x$extent, nsmall = 4), collapse = ", ")),
-              sprintf("crs           : %s\n", x$crs),
-              sprintf("values        : %s (min, max - range from entire extent)\n", paste(format(x$values, nsmall = 4), collapse = ", ")),
-              sprintf("window extent : %s\n", windowdescription), sprintf("window index  : %s\n", windowindex))
+  fmt <- list(sprintf("class            : %s\n", x$classname),
+              sprintf("dimensions       : %s (nrow, ncol)\n", paste(x$dimension[2:1], collapse = ", ")),
+              sprintf("resolution       : %s (x, y)\n", paste(format(x$resolution, nsmall = 4), collapse = ", ")),
+              sprintf("extent           : %s (xmin, xmax, ymin, ymax)\n", paste(format(x$extent, nsmall = 4), collapse = ", ")),
+              sprintf("crs              : %s\n", x$crs),
+              sprintf("values           : %s (min, max - range from entire extent)\n", paste(format(x$values, nsmall = 4), collapse = ", ")),
+              sprintf("window extent    : %s\n", windowdescription),
+              sprintf("window index     : %s\n", windowindex),
+              sprintf("window dimension : %s (ncol, nrow)\n", windowdim))
   fmt
 }